#56 refactor: DTO 폴더 내 response, request 폴더 생성 및 리팩토링
