+++ src/main/java/com/boaz/backend/domain/curriculum/dto/response/CurriculumStepResponse.java
@@ -1,4 +1,4 @@
-package com.boaz.backend.domain.curriculum.dto;
+package com.boaz.backend.domain.curriculum.dto.response;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.Getter;
