+++ src/main/java/com/boaz/backend/domain/recruitment/dto/response/RecruitmentStatusResponse.java
@@ -1,4 +1,4 @@
-package com.boaz.backend.domain.recruitment.dto;
+package com.boaz.backend.domain.recruitment.dto.response;
 
 import com.fasterxml.jackson.annotation.JsonInclude;
